moved two factor styledComponent util to component directory
